Change Color fields to u8

+++ src/color.rs
@@ -1,12 +1,12 @@
 #[derive(Clone)]
 pub struct Color {
-    pub R: u32,
-    pub G: u32,
-    pub B: u32,
+    pub R: u8,
+    pub G: u8,
+    pub B: u8,
 }
 
 impl Color {
-    pub fn new(R: u32, G: u32, B: u32) -> Self {
+    pub fn new(R: u8, G: u8, B: u8) -> Self {
         Color { R, G, B }
     }
 }
